requirements.txt
pandas 
Faker
google-cloud-bigquery